Fix Dialyzer Type Error
Since we don't explicit enforce any type in the implementation function
the parameter type is inferred from the first function that receive it's
value (at least is this case), since we pass it to Floki.find/2 the
html_node parameter typespec is binary() | Floki.html_node(), which
diverges from the behaviour typespec that is just Floki.html_node()
making dialyzer very angry with the project's code.
For now, I'm just going to update the callback typespec to reflect the
reality but IMO both callback and implementation should be limited to
accept only Floki.html_node() values.